pyproject.toml
[tool.poetry]
name = "notion-uploader"
version = "0.1.0"
description = "A CLI tool for uploading markdown files to Notion pages using the Notion API."
authors = ["Evan Hourigan"]
license = "MIT"
readme = "README.md"
packages = [{ include = "notion_uploader" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
notion-client = "^2.0.0"
python-dotenv = "^1.0.0"
rich = "^13.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.poetry.scripts]
notion-upload = "notion_uploader.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 